Roster View Headers & Contents Added

+++ ManagersDesktopApplication/globals.h
@@ -8,7 +8,7 @@ extern const QString ssTitleStr;
 
 typedef std::vector<std::vector<std::string>> vectorStr2D;
 typedef std::vector<std::string> vectorStr;
-typedef unsigned long long ull;
+typedef unsigned long long size_t;
 
 /* Our inputs are read as QStrings, however we need to
 *  input STD strings to many of our classes.
